forgot to mention in previous commit, moving forward, methods wont simply bail out of the jvm to shell anymore: attempting to do proper exception handling. FileWriter writer not initialized and in .close() no checking was done to see if it is assigned in the first place causing npe's. cleaner distinction between opening a new log file vs append.

+++ src/ca/vire/otp/LogFile.java
@@ -3,7 +3,7 @@ package ca.vire.otp;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.io.FileNotFoundException;
+
 import ca.vire.otp.utils.DateTime;
 
 /*
@@ -15,43 +15,65 @@ import ca.vire.otp.utils.DateTime;
 
 public final class LogFile {
 
-	private FileWriter writer;
+	private FileWriter writer = null;
 
-	public LogFile(String filename) {
-		try {
-			writer = new FileWriter(new File(filename), true);
-			writeHeader();
-		} catch (FileNotFoundException e) {
-			System.out.println(e.getMessage());
-		} catch (IOException e) {
-			e.printStackTrace();
+	public LogFile(String filename) throws IOException {
+		File log;
+
+		// Check if the log file already exists
+		log = new File(filename);
+
+		if (log.exists()) {
+			// Open for append
+			System.out.println("LogFile(): " + filename + " exists; appending");
+			try {
+				writer = new FileWriter(new File(filename), true);
+			} catch (IOException e) {
+				System.out.println("LogFile(): couldnt append to " + filename + ": " + e.getMessage());
+				e.printStackTrace();
+				throw e;
+			}
+		} else {
+			// or create new file
+			System.out.println("LogFile(): creating " + filename);
+			try {
+				writer = new FileWriter(new File(filename));
+			} catch (IOException e) {
+				System.out.println("LogFile(): couldnt open " + filename + ": " + e.getMessage());
+				e.printStackTrace();
+				throw e;
+			}
 		}
+		writeHeader();
 	}
 
-	public void writeChar(char c) {
+	public void writeChar(char c) throws IOException {
 		try {
 			writer.write(c);
 		} catch (IOException e) {
 			e.printStackTrace();
+			throw e;
 		}
 	}
 
 	public void close() {
-		try {
-			writer.close();
-		} catch (IOException e) {
-			e.printStackTrace();
+		if (writer != null) {
+			try {
+				writer.close();
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
 		}
 	}
 
-	private void writeHeader() {
-		//String timestamp = ca.vire.otp.utils.DateTime();
-		String timestamp = "";
+	private void writeHeader() throws IOException {
+		String timestamp = DateTime.getTimeStamp();
 
 		try {
 			writer.write("\n" + timestamp + "\n");
 		} catch (IOException e) {
 			e.printStackTrace();
+			throw e;
 		}
 	}
